Made it OAuth agonostic
Not everyone will use Vatsim Connect directly, but their own core login system that might not have the exact same data structure. This update allows to specify the data structure in the .env and makes it technically work with any OAuth2 provider. This will make it work with VATSCA's and VATUK's login systems to mentioned some.